requirements.txt
pandas
numpy
scikit-learn
seaborn
Flask
catboost
cassandra-driver